Corrected the division which now returns an integer

+++ ast/evaluator.py
@@ -20,7 +20,7 @@ class Evaluator:
         elif op == '-':
             return left - right
         elif op == '/':
-            return left / right
+            return int(left / right)
         elif op == '&':
             if not left:
                 return 0
